+++ Yet Another Simplifier/Tokens/VariableToken.cs
@@ -8,13 +8,13 @@ namespace Yet_Another_Simplifier.Tokens
 {
     public class VariableToken : Token
     {
-        public VariableToken(double quotient, List<Variable> variables)
+        public VariableToken(decimal quotient, List<Variable> variables)
         {
             Quotient = quotient;
             Variables = variables;
         }
 
-        public double Quotient { get; set; }
+        public decimal Quotient { get; set; }
         public List<Variable> Variables { get; set; }
 
         public override void NegateValue()
